refactor: move product selector to product folder
