+++ src/pages/Product/ProductDescription/ProductDescription.tsx
@@ -1,11 +1,11 @@
 import { ReactElement, memo } from 'react';
 import { useSelector } from 'react-redux';
+import { selectProduct } from '../../../store/product/selectors';
 import { splitToParagraphs } from '../../../utils';
 import { Accordion } from '../../../components/shared/Accordion';
 import { productAccordionData } from '../../../constant';
 
 import styles from './productDescription.module.scss';
-import { selectProduct } from '../../../store/cart/selectors';
 
 function ProductDescriptionComp(): ReactElement {
   const { product } = useSelector(selectProduct);
