+++ src/store/cart/selectors/index.ts
@@ -2,13 +2,11 @@ import { createSelector } from '@reduxjs/toolkit';
 import { RootState } from '../../store';
 import { TStatus } from '../../../types/types';
 import { ICart, ICartSlice, IDiscount } from '../types';
-import { IProductSlice } from '../../product/types';
 
 export const selectCart = (state: RootState): ICart => state.cart.basket;
 export const selectCartLoadingStatus = (state: RootState): TStatus => state.cart.status;
 export const selectCartError = (state: RootState): string => state.cart.error;
 export const selectDiscountsCodes = (state: RootState): IDiscount[] => state.cart.discounts;
-export const selectProduct = (state: RootState): IProductSlice => state.product;
 
 export const selectCartLoadingInfo = createSelector(
   selectCartLoadingStatus,
