+++ src/store/product/selectors/index.ts
@@ -1,0 +1,4 @@
+import { RootState } from '../../store';
+import { IProductSlice } from '../types';
+
+export const selectProduct = (state: RootState): IProductSlice => state.product;
